--- 2020-11/spring-03/spring-03-exercise/src/main/java/ru/otus/spring/service/IQuestionLine.java
@@ -1,16 +1,0 @@
-package ru.otus.spring.service;
-
-public interface IQuestionLine {
-    String getQuestionText();
-    void setQtext(String text);
-
-    boolean getAnswerCorrect();
-    void setAnswerCorrect(boolean bool);
-
-    public String getAnswerText();
-    void setAnswerText(String text);
-
-    public String getCorrectAnswerText();
-    void setCorrectAnswerText(String text);
-
-}
