+++ 2020-11/spring-03/spring-03-exercise/src/main/java/ru/otus/spring/service/QuestionLineImpl.java
@@ -1,0 +1,42 @@
+package ru.otus.spring.service;
+
+public class QuestionLineImpl implements QuestionLine {
+    private String qText;               // текст вопроса
+    private String answerText;          // текст ответа
+    private String correctAnswerText;   // текст правильного ответа
+    private boolean isAnswerCorrect;    // признак, является ли данный ответ правильным
+
+    public void setQtext(String text) {
+        this.qText = text;
+    }
+
+    public void setAnswerText(String answerText) {
+        this.answerText = answerText;
+    }
+
+    public void setCorrectAnswerText(String text) {
+        this.correctAnswerText = text;
+    }
+
+    public String getCorrectAnswerText() {
+        return correctAnswerText;
+    }
+
+    public void setAnswerCorrect(boolean bool) {
+        this.isAnswerCorrect = bool;
+    }
+
+    public String getQuestionText() {
+        return qText;
+    }
+
+    public String getAnswerText() {
+        return answerText;
+    }
+
+    public boolean getAnswerCorrect() {
+        return isAnswerCorrect;
+    }
+
+
+}
