+++ 2020-11/spring-09/spring-09-exercise/spring-shell.log
@@ -1,0 +1,52 @@
+1609777194243:help
+1609777197836:show
+1609777202416:exit
+1609779082958:help
+1609779088548:show
+1609779095415:exit
+1609782029544:help
+1609782033944:del 123
+1609782036955:del 1
+1609782040290:exit
+1609782350926:del 2
+1609782354324:del 1
+1609782360931:exit
+1609783798481:show
+1609783817522:exit
+1609784126534:show
+1609784129676:exit
+1609787597626:help
+1609787606718:ins 2 3 eerrtt
+1609787620755:show
+1609787630882:del 2
+1609787634547:del 3
+1609787642760:exit
+1609787755965:ins 2 3 rrttyy
+1609787764628:show
+1609787870333:ins 2 3 ddff
+1609787878207:show
+1609787978305:ins 2 3 rree
+1609787982737:show
+1609788018894:ins 1 2 eerr
+1609788170831:upd 2 3 rrr
+1609788188028:upd 1 2 3 err
+1609788193601:ins 2 3 rttyy
+1609788198444:exit
+1609788375959:ins 1 2 ddddddd
+1609788478839:del 1
+1609788788174:exit
+1609788949391:help
+1609788955827:ins 1 2 eerrttyy
+1609788960444:show
+1609788976999:upd 2 3 2 rrrrrrrrrrrrrrrrrrrrrrrrrrr
+1609788980332:show
+1609788984659:del 2
+1609788991966:exit
+1609789612964:help
+1609789627801:comments 1
+1609790418028:comments 2
+1609790424244:comments 1
+1609790440343:exit
+1609790581291:comments 2
+1609790587703:comments 1
+1609790594293:exit
